pyproject.toml
[project]
name = "django-starter-kit"
version = "0.1.0"
description = "A modern, high-performance Django 5.2 LTS starter kit with best practices"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2",
    "djangorestframework>=3.15.0",
    "django-environ>=0.11.2",
    "django-allauth[idp-oidc,socialaccount]>=65.11.0",
    "django-vite>=3.0.0",
    "django-imagekit>=5.0.0",
    "django-storages>=1.14.0",
    "django-anymail>=10.0",
    "django-actioncable>=1.0.4",
    "django-cors-headers>=4.3.0",
    "django-redis>=5.3.0",
    "whitenoise>=6.6.0",
    "celery>=5.3.0",
    "dj-database-url>=3.0.1",
    "granian[reload]>=2.5.0",
    "ruff>=0.12.0",
    "sentry-sdk>=1.40.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-django",
    "opentelemetry-instrumentation-celery",
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "psycopg>=3.2.9",
    "pillow>=10.0.0",
    "boto3>=1.30.0",
    "requests>=2.31.0",
    "amqp>=5.1.0",
    "channels-redis>=4.0.0",
    "django-debug-toolbar>=6.0.0",
    "pyjwt>=2.10.0",
    "cryptography>=45.0.6",
]
[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["."]
